+++ src/views/AccountDetails/AccountDetails.js
@@ -27,6 +27,7 @@ import {
 import {
   calculateTotalPaymentAmount,
   isRefundAllowed,
+  isCancelAllowed,
 } from '../../components/Accounts/accountFunctions';
 
 import css from './AccountDetails.css';
@@ -296,6 +297,7 @@ class AccountDetails extends React.Component {
 
     const totalPaidAmount = calculateTotalPaymentAmount(resources?.accounts?.records, feeFineActions);
     const refundAllowed = isRefundAllowed(account, feeFineActions);
+    const cancelAllowed = isCancelAllowed(account);
 
     return (
       <Paneset isRoot>
@@ -347,7 +349,7 @@ class AccountDetails extends React.Component {
               </Button>
               <Button
                 id="errorAccountActionsHistory"
-                disabled={disabled || buttonDisabled || isActionsPending || isAccountsPending}
+                disabled={disabled || buttonDisabled || isActionsPending || isAccountsPending || !cancelAllowed}
                 buttonStyle="primary"
                 onClick={this.error}
               >