+++ test/bigtest/network/config.js
@@ -147,6 +147,8 @@ export default function config() {
       const filterField = 'active';
       const departmentsField = 'departments';
       const usernameField = 'username';
+      const lastNameField = 'personal.lastName';
+      const barcodeField = 'barcode';
       if (/^%28/.test(query)) {
         query = decodeURIComponent(query);
       }
@@ -172,6 +174,16 @@ export default function config() {
       if (field === usernameField) {
         return users.where({
           [usernameField]: term.replace('*', '')
+        });
+      }
+      if (field === lastNameField) {
+        return users.where(u => {
+          return (u.personal.lastName === term.replace('*', ''));
+        });
+      }
+      if (field === barcodeField) {
+        return users.where({
+          [barcodeField]: term.replace('*', '')
         });
       }
     }